Fixing the bug
NSUserDefaults’s integerForKey() method returns 0 if it cannot find the value for the key you specify when reset Simulator, but in this app 0 is a valid row index. At this point the app doesn’t have any checklists yet, so index 0 does not exist in the lists array. That is why the app crashes. What you would like instead, is that NSUserDefaults returns -1 also if the “ChecklistIndex” key isn’t set, because to this app -1 means: show the main screen instead of a specific checklist.Here we specify default values for the default values of variable in NSUserDefaults.